Split `mineProductionRateEnumToString` out of `Common.h`
