+++ appOPHD/Common.h
@@ -2,6 +2,8 @@
 
 #include "EnumProductType.h"
 
+#include "MineProductionRateString.h"
+
 #include <NAS2D/Math/Rectangle.h>
 
 #include <map>
@@ -13,13 +15,10 @@ namespace NAS2D::Xml
 	class XmlDocument;
 }
 
-enum class MineProductionRate;
 enum class TerrainType;
 
 
 extern const std::map<TerrainType, std::string> TILE_INDEX_TRANSLATION;
-
-const std::string& mineProductionRateEnumToString(MineProductionRate mineProductionRate);
 
 void checkSavegameVersion(const std::string& filename);
 NAS2D::Xml::XmlDocument openSavegame(const std::string& filename);
