+++ appOPHD/MineProductionRateString.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+enum class MineProductionRate;
+
+
+const std::string& mineProductionRateEnumToString(MineProductionRate mineProductionRate);
