+++ appOPHD/appOPHD.vcxproj
@@ -206,6 +206,7 @@
     <ClCompile Include="MapObjects\Structures\SolarPlant.cpp" />
     <ClCompile Include="MapObjects\Structures\StorageTanks.cpp" />
     <ClCompile Include="MicroPather\micropather.cpp" />
+    <ClCompile Include="MineProductionRateString.cpp" />
     <ClCompile Include="MoraleString.cpp" />
     <ClCompile Include="ProductCatalogue.cpp" />
     <ClCompile Include="ProductPool.cpp" />
@@ -345,6 +346,7 @@
     <ClInclude Include="MapObjects\Structures\University.h" />
     <ClInclude Include="MapObjects\Structures\Warehouse.h" />
     <ClInclude Include="MicroPather\micropather.h" />
+    <ClInclude Include="MineProductionRateString.h" />
     <ClInclude Include="MoraleString.h" />
     <ClInclude Include="ProductCatalogue.h" />
     <ClInclude Include="ProductionCost.h" />
